Trait bytecode::access_path::AccessPathMap

pub trait AccessPathMap<T: AbstractDomain> {
    fn get_access_path(&self, ap: AccessPath) -> Option<&T>;

    fn remove_access_path(&mut self, ap: AccessPath) -> Option<T>;
}

Trait for a domain that can be viewed as a partial map from access paths to values and values can be deleted using their access paths

Required methods

fn get_access_path(&self, ap: AccessPath) -> Option<&T>

fn remove_access_path(&mut self, ap: AccessPath) -> Option<T>

Loading content...