Module bytecode::access_path_trie

The obvious approach to abstracting a set of concrete paths is using a set of abstract paths. An access path trie represents a set of paths in a way that avoids redundant representations of the same memory. Root nodes are access path roots and each internal node is an access path offset. Each node is (optionally) associated with abstract value of a generic type T.

Structs

AccessPathTrie
Set of (root node, child) associations

AccessPathTrieDisplay

TrieNode
A node in the access Trie: data associated with the parent node + children mapping offsets to child nodes