Struct bytecode::access_path_trie::AccessPathTrie

pub struct AccessPathTrie<T: FootprintDomain>(_);

Set of (root node, child) associations

Implementations

impl<T: FootprintDomain> AccessPathTrie<T>

pub fn remove_node(&mut self, ap: AccessPath) -> Option<TrieNode<T>>

Removes node located at the given access path Returns the node if it has been fully removed from the trie (i.e. it did not have any children)

pub fn get_child_data(&self) -> Option<T>

pub fn update_access_path_weak(&mut self, ap: AccessPath, data: Option<T>)

Like update_access_path, but always performs a weak update

pub fn update_access_path(&mut self, ap: AccessPath, data: Option<T>)

Update ap in global. Performs a strong update if the base of ap is a local and all offsets are Field’s. Otherwise, performs a weak update (TODO: more details). Creates nodes for each offset in ap if they do not already exist

pub fn join_access_path(&mut self, ap: AccessPath, node: TrieNode<T>)

Join the value bound to ap with node

pub fn bind_local(
&mut self,
local_index: TempIndex,
data: T,
fun_env: &FunctionEnv<'_>
)

Bind data to local_index in the trie, overwriting the old value of local_index

pub fn bind_local_node(
&mut self,
local_index: TempIndex,
node: TrieNode<T>,
fun_env: &FunctionEnv<'_>
)

Bind node to local_index in the trie, overwriting the old value of local_index

pub fn remove_local(&mut self, local_index: TempIndex, fun_env: &FunctionEnv<'_>)

Remove the value bound to the local variable local_index

pub fn bind_return(&mut self, return_index: usize, data: T)

Bind data to the return variable return_index

pub fn get_local(
&self,
local_index: TempIndex,
fun_env: &FunctionEnv<'_>
) -> Option<&T>

Retrieve the data associated with local_index in the trie. Returns None if there is no associated data

pub fn get_local_node(
&self,
local_index: TempIndex,
fun_env: &FunctionEnv<'_>
) -> Option<&TrieNode<T>>

Retrieve the node associated with local_index in the trie. Returns None if there is no associated node

pub fn local_exists(
&self,
local_index: TempIndex,
fun_env: &FunctionEnv<'_>
) -> bool

Return true if there is a value bound to local variable local_index

pub fn keys_statically_known(&self) -> bool

Return true if the keys of self have no dynamic components and thus can be converted into a compact set of concrete access paths.

pub fn substitute_footprint<F>(
self,
actuals: &[TempIndex],
type_actuals: &[Type],
func_env: &FunctionEnv<'_>,
sub_map: &dyn AccessPathMap<AbsAddr>,
sub_data: F
) -> Selfwhere
F: FnMut(&mut T, &[TempIndex], &[Type], &FunctionEnv<'_>, &dyn AccessPathMap<AbsAddr>) + Copy,

Bind caller data in actuals, type_actuals, and sub_map to self. (1) Bind all free type variables in self to type_actuals (2) Apply sub_data to self.data and (recursively) to the data fields of self.children

pub fn substitute_footprint_skip_data(
self,
actuals: &[TempIndex],
type_actuals: &[Type],
func_env: &FunctionEnv<'_>,
sub_map: &dyn AccessPathMap<AbsAddr>
) -> Self

Same as substitute_footprint, but does not change the data field of any node

pub fn substitute_footprint_concrete(
self,
actuals: &[TempIndex],
type_actuals: &[TypeTag],
func_env: &FunctionEnv<'_>,
sub_map: &dyn AccessPathMap<AbsAddr>,
env: &GlobalEnv
) -> Self

Substitute concrete values actuals and type_actuals into self

pub fn iter_values<F>(&mut self, f: F)where
F: FnMut(&mut TrieNode<T>) + Copy,

Apply f to each node in self

pub fn iter_offsets<F>(&self, f: F)where
F: FnMut(&Offset),

Apply f to each offset in self

pub fn iter_paths_opt<F>(&self, f: F)where
F: FnMut(&AccessPath, &Option<&T>),

Apply f to each (access path, Option(data)) pair encoded in self

pub fn iter_paths<F>(&self, f: F)where
F: FnMut(&AccessPath, &T),

Apply f to each (access path, data) pair encoded in self

pub fn filter_map_paths<F, R>(&self, f: F) -> Vec<R> where
F: FnMut(&AccessPath, &T) -> Option<R>,

Apply f to each (access path, data) pair encoded in self and collects the result when f returns Some(r)

pub fn display<'a>(
&'a self,
env: &'a FunctionEnv<'_>
) -> AccessPathTrieDisplay<'a, T>

Return a wrapper that of self that implements Display using env