Struct bytecode::function_data_builder::FunctionDataBuilder

pub struct FunctionDataBuilder<'env> {
    pub fun_env: &'env FunctionEnv<'env>,
    pub data: FunctionData,
    pub options: FunctionDataBuilderOptions,
    /* private fields */
}

A builder for FunctionData.

Fields
fun_env: &'env FunctionEnv<'env>data: FunctionDataoptions: FunctionDataBuilderOptions
Implementations

impl<'env> FunctionDataBuilder<'env>

pub fn new_with_options(
fun_env: &'env FunctionEnv<'env>,
data: FunctionData,
options: FunctionDataBuilderOptions
) -> Self

Creates a new builder with customized options

pub fn new(fun_env: &'env FunctionEnv<'env>, data: FunctionData) -> Self

Creates a new builder with options set to default values

pub fn get_target(&self) -> FunctionTarget<'_>

Gets a function target viewpoint on this builder. This locks the data for mutation until the returned value dies.

pub fn add_return(&mut self, ty: Type) -> usize

Add a return parameter.

pub fn set_loc_and_vc_info(&mut self, loc: Loc, message: &str)

Sets the default location as well as information about the verification condition message associated with the next instruction generated with emit_with.

pub fn set_loc_from_attr(&mut self, attr_id: AttrId)

Sets the default location from a code attribute id.

pub fn get_loc(&self, attr_id: AttrId) -> Loc

Gets the location from the bytecode attribute.

pub fn new_attr(&mut self) -> AttrId

Creates a new bytecode attribute id with default location.

pub fn new_label(&mut self) -> Label

Creates a new branching label for bytecode.

pub fn emit(&mut self, bc: Bytecode)

Emits a bytecode.

pub fn emit_vec(&mut self, bcs: Vec<Bytecode>)

Emits a sequence of bytecodes.

pub fn emit_with<F>(&mut self, f: F)where
F: FnOnce(AttrId) -> Bytecode,

Emits a bytecode via a function which takes a freshly generated attribute id.

pub fn set_next_debug_comment(&mut self, comment: String)

Sets the debug comment which should be associated with the next instruction emitted with self.emit_with(|id| ..).

pub fn clear_next_debug_comment(&mut self)

This will clear the state that the next self.emit_with(..) will add a debug comment.

pub fn emit_let(&mut self, def: Exp) -> (TempIndex, Exp)

Emits a let: this creates a new temporary and emits an assumption that this temporary is equal to the given expression. This can be used to abbreviate large expressions which are used multiple times, or get the value of an expression into a temporary for bytecode. Returns the temporary and a local expression referring to it.

pub fn emit_let_havoc(&mut self, ty: Type) -> (TempIndex, Exp)

Emits a new temporary with a havoced value of given type.