Module bytecode::function_target_pipeline

Structs

FunctionTargetPipeline
A processing pipeline for function targets.

FunctionTargetsHolder
A data structure which holds data for multiple function targets, and allows to manipulate them as part of a transformation pipeline.

ProcessorResultDisplay

Enums

FunctionVariant
Describes a function target variant.

VerificationFlavor
Describes a function verification flavor.

Traits

FunctionTargetProcessor
A trait describing a function target processor.