Module bytecode::local_mono_compat

A monomorphization processor for elimination of universal type quantifiers (forall t1: type, t2: type, ..., :: P<t1, t2, ...>) found in the expressions

This mono processor is a local monomorphization processor, it specializes a generic proposition with types found in the enclosing function only. This complements the global monomorphization processor in mono_analysis.rs which, if schedule to run after this local mono pass, instantiate axioms (and currently existential type quantifiers also) with information obtained in the whole GlobalEnv.

Local monomorphization can only be performed when two conditions hold:

all generic proposition are in the form of top-level universal type quantifiers, i.e., the proposition P must be in the form of forall t1: type, t2: type, ... : expr<t1, t2, ...> and there is no operation over P (e.g., !P, P ==> Q, etc. are not allowed).

the global invariant instrumentation pass places a global invariant I into all relevant instantiations of a function F. For example, if I<T> talks about a memory S<T> that is also going to be touched by F<T>, then I<T> must be instrumented in the generic version of F as well as in every instantiation of F (e.g., F<bool>, F<u64>, etc).

If both conditions hold, we can run local monomorphization, i.e., assert properties that are only relevant to the enclosing function only and safely assume that the properties hold for all other types that are not touched by the function.

TODO(mengxu) this is added to maintain backward compatibility while the implementation and porting of generic invariants are in progress.

Structs

LocalMonoCompatProcessor
This is the monomorphization processor that works on a function level.