Enum bytecode::stackless_bytecode::Bytecode

pub enum Bytecode {
    SpecBlock(AttrId, SpecBlockId),
    Assign(AttrId, TempIndex, TempIndex, AssignKind),
    Call(AttrId, Vec<TempIndex>, Operation, Vec<TempIndex>),
    Ret(AttrId, Vec<TempIndex>),
    Load(AttrId, TempIndex, Constant),
    Branch(AttrId, Label, Label, TempIndex),
    OnAbort(AttrId, Label, TempIndex),
    Jump(AttrId, Label),
    Label(AttrId, Label),
    Abort(AttrId, TempIndex),
    Nop(AttrId),
    SaveMem(AttrId, MemoryLabel, QualifiedId<StructId>),
    SaveSpecVar(AttrId, MemoryLabel, QualifiedId<SpecVarId>),
    Prop(AttrId, PropKind, Exp),
}

The stackless bytecode.

Variants

SpecBlock(AttrId, SpecBlockId)

Assign(AttrId, TempIndex, TempIndex, AssignKind)

Call(AttrId, Vec<TempIndex>, Operation, Vec<TempIndex>)

Ret(AttrId, Vec<TempIndex>)

Load(AttrId, TempIndex, Constant)

Branch(AttrId, Label, Label, TempIndex)

OnAbort(AttrId, Label, TempIndex)

Jump(AttrId, Label)

Label(AttrId, Label)

Abort(AttrId, TempIndex)

Nop(AttrId)

SaveMem(AttrId, MemoryLabel, QualifiedId<StructId>)

SaveSpecVar(AttrId, MemoryLabel, QualifiedId<SpecVarId>)

Prop(AttrId, PropKind, Exp)

Implementations

impl Bytecode

pub fn get_attr_id(&self) -> AttrId

pub fn is_exit(&self) -> bool

pub fn is_return(&self) -> bool

pub fn is_unconditional_branch(&self) -> bool

pub fn is_conditional_branch(&self) -> bool

pub fn is_branch(&self) -> bool

pub fn branch_dests(&self) -> Vec<Label>

Return the destination(s) if self is a branch/jump instruction

pub fn label_offsets(code: &[Bytecode]) -> BTreeMap<Label, CodeOffset>

Returns a mapping from labels to code offsets.

pub fn get_successors(
pc: CodeOffset,
code: &[Bytecode],
label_offsets: &BTreeMap<Label, CodeOffset>
) -> Vec<CodeOffset>

Return the successor offsets of this instruction. In addition to the code, a map of label to code offset need to be passed in.

pub fn get_exits(code: &[Bytecode]) -> Vec<CodeOffset>

Returns the code offsets at which the code exits(aborts or returns).

pub fn remap_all_vars<F>(
self,
func_target: &FunctionTarget<'_>,
f: &mut F
) -> Self where
F: FnMut(TempIndex) -> TempIndex,

Remaps variables in the instruction.

pub fn remap_src_vars<F>(
self,
func_target: &FunctionTarget<'_>,
f: &mut F
) -> Self where
F: FnMut(TempIndex) -> TempIndex,

Remaps variables in source position in the instruction.

pub fn modifies(&self, fun_target: &FunctionTarget<'_>) -> Vec<TempIndex>

Return the temporaries this instruction modifies. This includes references where the instruction can have effect on.

impl Bytecode

pub fn display<'env>(
&'env self,
func_target: &'env FunctionTarget<'env>
) -> BytecodeDisplay<'env>

Creates a format object for a bytecode in context of a function target.