Crate bytecode

Modules

annotations

borrow_analysis
Data flow analysis computing borrow information for preparation of memory_instrumentation.

clean_and_optimize

dataflow_analysis
Adapted from AbstractInterpreter for Bytecode, this module defines the data-flow analysis framework for stackless bytecode.

eliminate_imm_refs

eliminate_mut_refs

function_target

function_target_pipeline

graph

livevar_analysis

memory_instrumentation

reaching_def_analysis

stackless_bytecode

stackless_bytecode_generator

stackless_control_flow_graph
Adapted from control_flow_graph for Bytecode, this module defines the control-flow graph on Stackless Bytecode used in analysis as part of Move prover.

test_instrumenter

usage_analysis

Functions

print_targets_for_test
Print function targets for testing and debugging.