Type Definition bytecode::access_path::AbsAddr

type AbsAddr = SetDomain<Addr>;

Abstraction of an address: non-empty set of constant or footprint address values

Implementations

impl AbsAddr

pub fn constant(a: BigUint) -> Self

Create a constant address from concrete address a

pub fn footprint(ap: AccessPath) -> Self

Create a footprint address from access path ap

pub fn formal(formal_index: TempIndex, func_env: &FunctionEnv<'_>) -> Self

Create a footprint address read from formal temp_index

pub fn is_constant(&self) -> bool

Return true if self is a constant

pub fn is_statically_known(&self) -> bool

Return true if self consists only of statically known constants

pub fn substitute_footprint(
&mut self,
actuals: &[TempIndex],
type_actuals: &[Type],
func_env: &FunctionEnv<'_>,
sub_map: &dyn AccessPathMap<AbsAddr>
)

Substitute all occurences of Footprint(ap) in self by resolving the accesss path ap in sub_map

pub fn add_struct_offset(
self,
mid: &ModuleId,
sid: StructId,
types: Vec<Type>
) -> Self

Return a new abstract address by adding the offset mid::sid<types> to each element of self

pub fn add_offset(&self, offset: Offset) -> Self

Return a new abstract address by adding the offset offset to each element of self

pub fn prepend(self, prefix: AccessPath) -> Self

Produce a new version of self with prefix prepended to each footprint value

pub fn footprint_paths(&self) -> impl Iterator<Item = &AccessPath>

return an iterator over the footprint paths in self

pub fn get_concrete_addresses(&self) -> Vec<AccountAddress>

Return an iterator over the concrete addresses in self

pub fn display<'a>(&'a self, env: &'a FunctionEnv<'_>) -> AbsAddrDisplay<'a>

Return a wrapper of self that implements Display using env