Struct bytecode::access_path_trie::AccessPathTrie

pub struct AccessPathTrie<T: FootprintDomain>(_);

Set of (root node, child) associations

Implementations

impl<T: FootprintDomain> AccessPathTrie<T>

pub fn update_access_path_weak(&mut self, ap: AccessPath, data: Option<T>)

Like update_access_path, but always performs a weak update

pub fn update_access_path(&mut self, ap: AccessPath, data: Option<T>)

Update ap in global. Performs a strong update if the base of ap is a local and all offsets are Field's. Otherwise, performs a weak update (TODO: more details). Creates nodes for each offset in ap if they do not already exist

pub fn join_access_path(&mut self, ap: AccessPath, node: TrieNode<T>)

Join the value bound to ap with node

pub fn bind_local(&mut self, local_index: TempIndex, data: T)

Bind data to local_index in the trie, overwriting the old value of local_index

pub fn bind_local_node(&mut self, local_index: TempIndex, node: TrieNode<T>)

Bind node to local_index in the trie, overwriting the old value of local_index

pub fn remove_local(&mut self, local_index: TempIndex)

Remove the value bound to the local variable local_index

pub fn bind_return(&mut self, return_index: usize, data: T)

Bind data to the return variable return_index

pub fn get_local(&self, local_index: TempIndex) -> Option<&T>

Retrieve the data associated with local_index in the trie. Returns None if there is no associated data

pub fn get_local_node(&self, local_index: TempIndex) -> Option<&TrieNode<T>>

Retrieve the node associated with local_index in the trie. Returns None if there is no associated node

pub fn local_exists(&self, local_index: TempIndex) -> bool

Return true if there is a value bound to local variable local_index

pub fn substitute_footprint<F>(
self,
actuals: &[AbsAddr],
type_actuals: &[Type],
sub_map: &dyn AccessPathMap<AbsAddr>,
sub_data: F
) -> Self where
F: FnMut(&mut T, &[AbsAddr], &[Type], &dyn AccessPathMap<AbsAddr>) + Copy,

Bind caller data in actuals, type_actuals, and sub_map to self. (1) Bind all free type variables in self to type_actuals (2) Apply sub_data to self.data and (recursively) to the data fields of self.children

pub fn substitute_footprint_skip_data(
self,
actuals: &[AbsAddr],
type_actuals: &[Type],
sub_map: &dyn AccessPathMap<AbsAddr>
) -> Self

Same as substitute_footprint, but does not change the data field of any node

pub fn iter_values<F>(&mut self, f: F) where
F: FnMut(&mut TrieNode<T>) + Copy,

Apply f to each node in self

pub fn iter_paths_opt<F>(&self, f: F) where
F: FnMut(&AccessPath, &Option<&T>),

Apply f to each (access path, Option(data)) pair encoded in self

pub fn iter_paths<F>(&self, f: F) where
F: FnMut(&AccessPath, &T),

Apply f to each (access path, data) pair encoded in self

pub fn display<'a>(
&'a self,
env: &'a FunctionTarget<'_>
) -> AccessPathTrieDisplay<'a, T>

Return a wrapper that of self that implements Display using env