Trait bytecode::compositional_analysis::CompositionalAnalysis

pub trait CompositionalAnalysis<Domain: AbstractDomain + 'static>: DataflowAnalysis where
    Self::State: AbstractDomain + 'static, {
    fn to_summary(&self, state: Self::State) -> Domain;

    fn summarize(
        &self, 
        func_env: &FunctionEnv<'_>, 
        initial_state: Self::State, 
        data: FunctionData
    ) -> FunctionData { ... }
}

Trait that lifts an intraprocedural analysis into a bottom-up, compositional interprocedural analysis.

Required methods

fn to_summary(&self, state: Self::State) -> Domain

Specifies mapping from elements of dataflow analysis domain to elements of Domain.

Loading content...
Provided methods

fn summarize(
&self,
func_env: &FunctionEnv<'_>,
initial_state: Self::State,
data: FunctionData
) -> FunctionData

Computes a postcondition for the procedure data and then maps the postcondition to an element of abstract domain Domain by applying to_summary function. The result is stored in the summary cache of data.

Loading content...