Struct bytecode::function_data_builder::FunctionDataBuilder

pub struct FunctionDataBuilder<'env> {
    pub fun_env: &'env FunctionEnv<'env>,
    pub data: FunctionData,
    // some fields omitted
}

A builder for FunctionData.

Fields
fun_env: &'env FunctionEnv<'env>data: FunctionData
Implementations

impl<'env> FunctionDataBuilder<'env>

pub fn new(fun_env: &'env FunctionEnv<'env>, data: FunctionData) -> Self

Creates a new builder.

pub fn global_env(&self) -> &'env GlobalEnv

Gets the global env associated with this builder.

pub fn get_target(&self) -> FunctionTarget<'_>

Gets a function target viewpoint on this builder. This locks the data for mutation until the returned value dies.

pub fn new_temp(&mut self, ty: Type) -> TempIndex

Allocates a new temporary.

pub fn add_return(&mut self, ty: Type) -> usize

Add a return parameter.

pub fn set_loc(&mut self, loc: Loc)

Sets the default location.

pub fn set_loc_and_vc_info(&mut self, loc: Loc, message: &str)

Sets the default location as well as information about the verification condition message associated with the next instruction generated with emit_with.

pub fn set_loc_from_attr(&mut self, attr_id: AttrId)

Sets the default location from a code attribute id.

pub fn set_loc_from_node(&mut self, node_id: NodeId)

Sets the default location from a node id.

pub fn get_loc(&self, attr_id: AttrId) -> Loc

Gets the location from the bytecode attribute.

pub fn get_current_loc(&self) -> &Loc

Gets the default location.

pub fn new_attr(&mut self) -> AttrId

Creates a new bytecode attribute id with default location.

pub fn new_label(&mut self) -> Label

Creates a new branching label for bytecode.

pub fn new_node(&self, ty: Type, inst_opt: Option<Vec<Type>>) -> NodeId

Creates a new expression node id, using current default location, provided type, and optional instantiation.

pub fn mk_bool_const(&self, value: bool) -> Exp

Make a boolean constant expression.

pub fn mk_call(&self, ty: &Type, oper: Operation, args: Vec<Exp>) -> Exp

Makes a Call expression.

pub fn mk_ite(&self, cond: Exp, if_true: Exp, if_false: Exp) -> Exp

Makes an if-then-else expression.

pub fn mk_bool_call(&self, oper: Operation, args: Vec<Exp>) -> Exp

Makes a Call expression with boolean result type.

pub fn mk_not(&self, arg: Exp) -> Exp

Make a boolean not expression.

pub fn mk_eq(&self, arg1: Exp, arg2: Exp) -> Exp

Make an equality expression.

pub fn mk_and(&self, arg1: Exp, arg2: Exp) -> Exp

Make an and expression.

pub fn mk_or(&self, arg1: Exp, arg2: Exp) -> Exp

Make an or expression.

pub fn mk_implies(&self, arg1: Exp, arg2: Exp) -> Exp

Make an implies expression.

pub fn mk_builtin_num_const(&self, oper: Operation) -> Exp

Make a numerical expression for some of the builtin constants.

pub fn mk_join_bool(
&self,
oper: Operation,
args: impl Iterator<Item = Exp>
) -> Option<Exp>

Join an iterator of boolean expressions with a boolean binary operator.

pub fn mk_join_opt_bool(
&self,
oper: Operation,
arg1: Option<Exp>,
arg2: Option<Exp>
) -> Option<Exp>

Join two boolean optional expression with binary operator.

pub fn mk_vector_quant_opt<F>(
&self,
kind: QuantKind,
vector: Exp,
elem_ty: &Type,
f: &mut F
) -> Option<Exp> where
F: FnMut(Exp) -> Option<Exp>,

Creates a quantifier over the content of a vector. The passed function f receives an expression representing an element of the vector and returns the quantifiers predicate; if it returns None, this function will also return None, otherwise the quantifier will be returned.

pub fn mk_mem_quant_opt<F>(
&self,
kind: QuantKind,
mem: QualifiedId<StructId>,
f: &mut F
) -> Option<Exp> where
F: FnMut(Exp) -> Option<Exp>,

Creates a quantifier over the content of memory. The passed function f receives

pub fn mk_decl(
&self,
name: Symbol,
ty: Type,
binding: Option<Exp>
) -> LocalVarDecl

Makes a local variable declaration.

pub fn mk_symbol(&self, str: &str) -> Symbol

Makes a symbol from a string.

pub fn mk_type_domain(&self, ty: Type) -> Exp

Makes a type domain expression.

pub fn mk_field_select(
&self,
field_env: &FieldEnv<'_>,
targs: &[Type],
exp: Exp
) -> Exp

Makes an expression which selects a field from a struct.

pub fn mk_temporary(&self, temp: TempIndex) -> Exp

Makes an expression for a temporary.

pub fn mk_local(&self, name: &str, ty: Type) -> Exp

Makes an expression for a named local.

pub fn get_memory_of_node(&self, node_id: NodeId) -> QualifiedId<StructId>

Get's the memory associated with a Call(Global,..) or Call(Exists, ..) node. Crashes if the the node is not typed as expected.

pub fn emit(&mut self, bc: Bytecode)

Emits a bytecode.

pub fn emit_vec(&mut self, bcs: Vec<Bytecode>)

Emits a sequence of bytecodes.

pub fn emit_with<F>(&mut self, f: F) where
F: FnOnce(AttrId) -> Bytecode,

Emits a bytecode via a function which takes a freshly generated attribute id.

pub fn set_next_debug_comment(&mut self, comment: String)

Sets the debug comment which should be associated with the next instruction emitted with self.emit_with(|id| ..).

pub fn clear_next_debug_comment(&mut self)

This will clear the state that the next self.emit_with(..) will add a debug comment.

pub fn emit_let(&mut self, def: Exp) -> (TempIndex, Exp)

Emits a let: this creates a new temporary and emits an assumption that this temporary is equal to the given expression. This can be used to abbreviate large expressions which are used multiple times, or get the value of an expression into a temporary for bytecode. Returns the temporary and a local expression referring to it.

pub fn emit_let_havoc(&mut self, ty: Type) -> (TempIndex, Exp)

Emits a new temporary with a havoced value of given type.