Struct bytecode::function_target::FunctionData

pub struct FunctionData {
    pub code: Vec<Bytecode>,
    pub local_types: Vec<Type>,
    pub return_types: Vec<Type>,
    pub param_proxy_map: BTreeMap<usize, usize>,
    pub ref_param_proxy_map: BTreeMap<usize, usize>,
    pub ref_param_return_map: BTreeMap<usize, usize>,
    pub acquires_global_resources: Vec<StructId>,
    pub locations: BTreeMap<AttrId, Loc>,
    pub annotations: Annotations,
    pub spec_blocks_on_impl: BTreeMap<SpecBlockId, CodeOffset>,
    pub name_to_index: BTreeMap<Symbol, usize>,
    pub modify_targets: BTreeMap<QualifiedId<StructId>, Vec<Exp>>,
}

Holds the owned data belonging to a FunctionTarget, contained in a FunctionTargetsHolder.

Fields
code: Vec<Bytecode>

The bytecode.

local_types: Vec<Type>

The locals, including parameters.

return_types: Vec<Type>

The return types.

param_proxy_map: BTreeMap<usize, usize>

TODO(wrwg): document what this is for

ref_param_proxy_map: BTreeMap<usize, usize>

A map from mut ref input parameters to the generated output parameters.

ref_param_return_map: BTreeMap<usize, usize>

A map from mut ref output parameters to the input parameters.

acquires_global_resources: Vec<StructId>

The set of global resources acquired by this function.

locations: BTreeMap<AttrId, Loc>

A map from byte code attribute to source code location.

annotations: Annotations

Annotations associated with this function.

spec_blocks_on_impl: BTreeMap<SpecBlockId, CodeOffset>

Map of spec block ids as given by the source, to the code offset in the original bytecode. Those spec block's content is found at func_env.get_specification_on_impl(code_offset).

name_to_index: BTreeMap<Symbol, usize>

A map from local names to temp indices in code.

modify_targets: BTreeMap<QualifiedId<StructId>, Vec<Exp>>

A cache of targets modified by this function.

Implementations

impl FunctionData

pub fn new(
func_env: &FunctionEnv<'_>,
code: Vec<Bytecode>,
local_types: Vec<Type>,
return_types: Vec<Type>,
locations: BTreeMap<AttrId, Loc>,
acquires_global_resources: Vec<StructId>,
given_spec_blocks: BTreeMap<SpecBlockId, CodeOffset>
) -> Self

Creates new function target data.

pub fn next_free_attr_index(&self) -> usize

Computes the next available index for AttrId.

pub fn next_free_label_index(&self) -> usize

Computes the next available index for Label.

pub fn get_callees(&self) -> BTreeSet<QualifiedId<FunId>>

Return the set of callees invoked by this function, including native functions

pub fn rename_vars<F>(&mut self, f: &F) where
F: Fn(TempIndex) -> TempIndex,

Apply a variable renaming to this data, adjusting internal data structures.