Trait bytecode::function_target_pipeline::FunctionTargetProcessor

pub trait FunctionTargetProcessor {
    fn process(
        &self, 
        targets: &mut FunctionTargetsHolder, 
        func_env: &FunctionEnv<'_>, 
        data: FunctionData
    ) -> FunctionData;

    fn name(&self) -> String;
}

A trait for processing a function target. Takes as parameter a target holder which can be mutated, the env of the function being processed, and the target data. During the time the processor is called, the target data is removed from the holder, and added back once transformation has finished. This allows the processor to take ownership on the target data. Notice that you can use FunctionTarget{func_env, &data} to temporarily construct a function target for access to the underlying data.

Required methods

fn process(
&self,
targets: &mut FunctionTargetsHolder,
func_env: &FunctionEnv<'_>,
data: FunctionData
) -> FunctionData

fn name(&self) -> String

Returns a name for this processor. This should be suitable as a file suffix.

Loading content...