Trait bytecode::function_target_pipeline::FunctionTargetProcessor

pub trait FunctionTargetProcessor {
    fn name(&self) -> String;

    fn process(
        &self, 
        _targets: &mut FunctionTargetsHolder, 
        _fun_env: &FunctionEnv<'_>, 
        _data: FunctionData
    ) -> FunctionData { ... }

    fn process_and_maybe_remove(
        &self, 
        targets: &mut FunctionTargetsHolder, 
        func_env: &FunctionEnv<'_>, 
        data: FunctionData
    ) -> Option<FunctionData> { ... }

    fn initialize(&self, _env: &GlobalEnv, _targets: &mut FunctionTargetsHolder) { ... }

    fn finalize(&self, _env: &GlobalEnv, _targets: &mut FunctionTargetsHolder) { ... }
}

A trait describing a function target processor.

Required methods

fn name(&self) -> String

Returns a name for this processor. This should be suitable as a file suffix.

Loading content...
Provided methods

fn process(
&self,
_targets: &mut FunctionTargetsHolder,
_fun_env: &FunctionEnv<'_>,
_data: FunctionData
) -> FunctionData

Processes a function variant. Takes as parameter a target holder which can be mutated, the env of the function being processed, and the target data. During the time the processor is called, the target data is removed from the holder, and added back once transformation has finished. This allows the processor to take ownership on the target data.

fn process_and_maybe_remove(
&self,
targets: &mut FunctionTargetsHolder,
func_env: &FunctionEnv<'_>,
data: FunctionData
) -> Option<FunctionData>

Same as process but can return None to indicate that the function variant is removed. By default, this maps to Some(self.process(..)). One needs to implement either this function or process.

fn initialize(&self, _env: &GlobalEnv, _targets: &mut FunctionTargetsHolder)

A function which is called once before any process call is issued.

fn finalize(&self, _env: &GlobalEnv, _targets: &mut FunctionTargetsHolder)

A function which is called once after the last process call.

Loading content...