Enum bytecode::stackless_bytecode::Bytecode

pub enum Bytecode {
    Assign(AttrId, TempIndex, TempIndex, AssignKind),
    Call(AttrId, Vec<TempIndex>, Operation, Vec<TempIndex>, Option<AbortAction>),
    Ret(AttrId, Vec<TempIndex>),
    Load(AttrId, TempIndex, Constant),
    Branch(AttrId, Label, Label, TempIndex),
    Jump(AttrId, Label),
    Label(AttrId, Label),
    Abort(AttrId, TempIndex),
    Nop(AttrId),
    SaveMem(AttrId, MemoryLabel, QualifiedInstId<StructId>),
    SaveSpecVar(AttrId, MemoryLabel, QualifiedInstId<SpecVarId>),
    Prop(AttrId, PropKind, Exp),
}

The stackless bytecode.

Variants

Assign(AttrId, TempIndex, TempIndex, AssignKind)

Tuple Fields of Assign

0: AttrId1: TempIndex2: TempIndex3: AssignKind

Call(AttrId, Vec<TempIndex>, Operation, Vec<TempIndex>, Option<AbortAction>)

Tuple Fields of Call

0: AttrId1: Vec<TempIndex>2: Operation3: Vec<TempIndex>4: Option<AbortAction>

Ret(AttrId, Vec<TempIndex>)

Tuple Fields of Ret

0: AttrId1: Vec<TempIndex>

Load(AttrId, TempIndex, Constant)

Tuple Fields of Load

0: AttrId1: TempIndex2: Constant

Branch(AttrId, Label, Label, TempIndex)

Tuple Fields of Branch

0: AttrId1: Label2: Label3: TempIndex

Jump(AttrId, Label)

Tuple Fields of Jump

0: AttrId1: Label

Label(AttrId, Label)

Tuple Fields of Label

0: AttrId1: Label

Abort(AttrId, TempIndex)

Tuple Fields of Abort

0: AttrId1: TempIndex

Nop(AttrId)

Tuple Fields of Nop

0: AttrId

SaveMem(AttrId, MemoryLabel, QualifiedInstId<StructId>)

Tuple Fields of SaveMem

0: AttrId1: MemoryLabel2: QualifiedInstId<StructId>

SaveSpecVar(AttrId, MemoryLabel, QualifiedInstId<SpecVarId>)

Tuple Fields of SaveSpecVar

0: AttrId1: MemoryLabel2: QualifiedInstId<SpecVarId>

Prop(AttrId, PropKind, Exp)

Tuple Fields of Prop

0: AttrId1: PropKind2: Exp

Implementations

impl Bytecode

pub fn get_attr_id(&self) -> AttrId

pub fn is_exit(&self) -> bool

pub fn is_return(&self) -> bool

pub fn is_unconditional_branch(&self) -> bool

pub fn is_conditional_branch(&self) -> bool

pub fn is_branch(&self) -> bool

pub fn branch_dests(&self) -> Vec<Label>

Return the destination(s) if self is a branch/jump instruction

pub fn label_offsets(code: &[Bytecode]) -> BTreeMap<Label, CodeOffset>

Returns a mapping from labels to code offsets.

pub fn get_successors(
pc: CodeOffset,
code: &[Bytecode],
label_offsets: &BTreeMap<Label, CodeOffset>
) -> Vec<CodeOffset>

Return the successor offsets of this instruction. In addition to the code, a map of label to code offset need to be passed in.

pub fn get_exits(code: &[Bytecode]) -> Vec<CodeOffset>

Returns the code offsets at which the code exits(aborts or returns).

pub fn remap_all_vars<F>(
self,
func_target: &FunctionTarget<'_>,
f: &mut F
) -> Self where
F: FnMut(TempIndex) -> TempIndex,

Remaps variables in the instruction.

pub fn remap_src_vars<F>(
self,
func_target: &FunctionTarget<'_>,
f: &mut F
) -> Self where
F: FnMut(TempIndex) -> TempIndex,

Remaps variables in source position in the instruction.

pub fn instantiate(&self, env: &GlobalEnv, params: &[Type]) -> Self

pub fn modifies(
&self,
func_target: &FunctionTarget<'_>
) -> (Vec<TempIndex>, Vec<(TempIndex, bool)>)

Return the temporaries this instruction modifies and how the temporaries are modified.

For a temporary with TempIndex $t, if $t is modified by the instruction and

$t is a value or an immutable reference, it will show up in the first Vec

$t is a mutable reference and only its value is modified, not the reference itself, it will show up in the second Vec as ($t, false).

$t is a mutable reference and the reference itself is modified (i.e., the location and path it is pointing to), it will show up in the second Vec as ($t, true).

impl Bytecode

pub fn display<'env>(
&'env self,
func_target: &'env FunctionTarget<'env>,
label_offsets: &'env BTreeMap<Label, CodeOffset>
) -> BytecodeDisplay<'env>

Creates a format object for a bytecode in context of a function target.